Refactoring

+++ src/camera/playerone/shim.rs
@@ -96,22 +96,22 @@ impl Camera {
 
     pub fn gain(&self) -> Result<(i64, bool), String> {
         self.get_config(POAConfig::POA_GAIN)
-            .map(|(v, auto)| (unsafe { v.intValue } as i64, auto))
+            .map(|(v, auto)| (unsafe { v.intValue }, auto))
     }
 
     pub fn offset(&self) -> Result<i64, String> {
         self.get_config(POAConfig::POA_OFFSET)
-            .map(|(v, _)| unsafe { v.intValue } as i64)
+            .map(|(v, _)| unsafe { v.intValue })
     }
 
     pub fn exposure(&self) -> Result<(i64, bool), String> {
         self.get_config(POAConfig::POA_EXPOSURE)
-            .map(|(v, auto)| (unsafe { v.intValue } as i64, auto))
+            .map(|(v, auto)| (unsafe { v.intValue }, auto))
     }
 
     pub fn cooler_power(&self) -> Result<i64, String> {
         self.get_config(POAConfig::POA_COOLER_POWER)
-            .map(|(v, _)| unsafe { v.intValue } as i64)
+            .map(|(v, _)| unsafe { v.intValue })
     }
 
     pub fn cooler(&self) -> Result<bool, String> {
@@ -121,7 +121,7 @@ impl Camera {
 
     pub fn dew_heater_power(&self) -> Result<i64, String> {
         self.get_config(POAConfig::POA_HEATER_POWER)
-            .map(|(v, _)| unsafe { v.intValue } as i64)
+            .map(|(v, _)| unsafe { v.intValue })
     }
 
     pub fn is_config_supported(&self, conf_id: POAConfig) -> bool {
@@ -235,7 +235,11 @@ impl Camera {
 
     pub fn start_exposure(&mut self, is_single: bool) -> Result<(), String> {
         let sdk = PlayerOneSdk::try_load().unwrap();
-        let flag = if is_single { POABool::POA_TRUE } else { POABool::POA_FALSE };
+        let flag = if is_single {
+            POABool::POA_TRUE
+        } else {
+            POABool::POA_FALSE
+        };
         let err = unsafe { sdk.api.POAStartExposure(self.id, flag) };
         if err == POAErrors::POA_OK {
             Ok(())
